+++ backend/src/routes/userRoute.ts
@@ -1,11 +1,9 @@
 import express from "express";
-import { createUser, getUserById, getUsers, } from "../controllers/userController";
+import { createUser, getUserById } from "../controllers/user";
 
 const router = express.Router();
 
-
-router.post("/users", createUser); 
-router.get("/users", getUsers); 
-router.get("/users/:id", getUserById); 
+router.post("/users", createUser);
+router.get("/users/:id", getUserById);
 
 export default router;
